Avoid zombies on androidscreencast
[Cherry-picked from ICS to JB]
When using a third party screen capture program called androidscreencast
(http://code.google.com/p/androidscreencast/) to get the framebuffer of
the device, there are tons of screencap zombie processes got left behind.
The issue is also mentioned here:
http://code.google.com/p/android/issues/detail?id=22836.

The cause of the issue is that adbd spawns off screencap processes,
and these child processes were not waited to be finished.
This change fixes the issue.

Change-Id: I4596b6e4d1d1dd9d76896fe79138dee4084d8406
Signed-off-by: Yuriy Zabroda <yuriy.zabroda@ti.com>